Convert Mana stat to a float

+++ src/main/java/com/chaosbuffalo/mkultra/core/IPlayerData.java
@@ -51,11 +51,11 @@ public interface IPlayerData {
 
     float getManaRegenRate();
 
-    void setMana(int mana);
+    void setMana(float mana);
 
-    int getMana();
+    float getMana();
 
-    default boolean consumeMana(int amount) {
+    default boolean consumeMana(float amount) {
         if (getMana() >= amount) {
             setMana(getMana() - amount);
             return true;
