+++ src/main/java/com/chaosbuffalo/mkultra/core/abilities/FlameWave.java
@@ -40,7 +40,7 @@ public class FlameWave extends PlayerAbility {
     }
 
     @Override
-    public int getManaCost(int currentRank) {
+    public float getManaCost(int currentRank) {
         return 8 + currentRank * 2;
     }
